pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "jq>=1.10.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "llama-index>=0.13.4",
    "llama-index-embeddings-huggingface>=0.6.0",
    "llama-index-vector-stores-chroma>=0.5.2",
    "python-dotenv>=1.1.1",
    "langchain-tavily>=0.2.11",
    "ipython"
]
